finished authorization

+++ YelpCamp/routes/campgrounds.js
@@ -1,21 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('./../utils/catchAsync');
-const ExpressError = require('./../utils/ExpressError');
 const Campground = require('./../models/campground');
-const {campgroundSchema} = require('./../schemas.js');
 const {isLoggedIn} = require('./../middleware')
+const {isAuthor} = require('./../middleware')
+const {validateCampground} = require('./../middleware')
 
-///////////////////JOI SERVER SIDE DATA VALIDATION MIDDLEWARE
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
 
 router.get('/', async (req, res) => {
     const campgrounds = await Campground.find({});
@@ -39,7 +29,12 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 }))
 
 router.get('/:id', catchAsync(async (req, res) => {
-    const campground = await (await Campground.findById(req.params.id).populate("reviews").populate("author"));
+    const campground = await (await Campground.findById(req.params.id).populate({
+        path: "reviews",
+        populate: {
+            path: "author"      //nested populate (how to populate 2 things: reviews and reviews author)
+        }
+        }).populate("author"));
     // console.log(campground)
     if(!campground){
         req.flash("error", "Cannot find that campground")
@@ -49,8 +44,9 @@ router.get('/:id', catchAsync(async (req, res) => {
     
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
     if(!campground){
         req.flash("error", "Cannot find that campground")
         return res.redirect('/campgrounds')
@@ -59,15 +55,20 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
 }))
 
 //                          OUR VALIDATION MIDDLEWARE
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    const c = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash("success", "Successfully updated campgound");
-    res.redirect(`/campgrounds/${campground._id}`);
+    res.redirect(`/campgrounds/${c._id}`);
 }))
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.author.equals(req.user._id)){
+        req.flash("error", "You do not have permission to do that");
+        return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash("success", "Successfully delete your campground");
     res.redirect('/campgrounds')
